chore(trien-khai): don cau hinh, kieu du lieu tien va README truoc khi len mang
Buoi 1 cua ke hoach trien khai - chin viec, khong viec nao doi hanh vi nguoi
dung nhin thay, nhung thieu chung thi ban len mang chet khong ro ly do.

CAU HINH (8.2.1, 8.2.3, 8.2.4, 8.2.5)

Doi chieu 156 bien ma nguon doc voi .env.example: thieu ba bien that
(GEMINI_CA_BUNDLE, SANCTUM_EXPIRATION_MINUTES, SANCTUM_STATEFUL_DOMAINS) va
thua VITE_APP_NAME - di san cua khung Laravel+Vite, du an nay dung Next.js
nen khong ma nao doc no.

Them canh bao TU KEU khi ban da trien khai chay voi cau hinh chi hop cho may
phat trien: APP_DEBUG bat, CORS mo cho moi ten mien, APP_KEY rong. Ba thu do
KHONG lam hong gi ngay - ung dung chay binh thuong, khong bao loi nao - nen
rat de len mang roi van con nguyen. Moi cai la mot cua mo: APP_DEBUG=true thi
mot loi 500 bat ky la in ra toan bo bien moi truong, ke ca khoa cong thanh
toan. Chi ghi log, khong chan khoi dong: dung duoc nhung khong vao duoc thi
con te hon, nhat la trong buoi bao ve.

KIEU DU LIEU TIEN (8.3.3)

Order, OrderDetail, OrderDetailTopping ep kieu tien la 'float' trong khi
Item::base_price va ItemPrice::price da la 'integer' - cung mot dong tien ma
hai kieu khac nhau tuy cho doc. Dong Viet Nam khong co don vi nho hon dong.

Kem lenh db:normalize-money don 534 o da lo ghi dang so thuc. Lenh MAC DINH
CHI BAO CAO, phai go --apply moi ghi - day la du lieu tien. Va no TU CHOI doi
bat ky o nao co phan thap phan: o nhu vay nghia la co mot cho tinh tien dang
sinh ra so le, cat di la giau loi va lam sai so tien.

Da chay tren du lieu that: doanh thu truoc va sau khop tung dong (172 hoa don,
45.116.000d ban hang, 27.579.500d goi).

DU LIEU DEMO (8.3.1, 8.3.4, 8.3.6)

db:indexes: 23 chi muc du, khong loi. DemoSeeder gieo tu CSDL TRONG chay sach
(kiem tren mot CSDL nhap rieng, khong dung du lieu that) va moi truong tien ra
so nguyen ngay tu dau - tuc ban va o tren da chan duoc goc re, khong chi don
hau qua.

Xoa mot tai khoan thu nghiem lac ("Perf Test" / Cafe Test Shot, tao 28/07 luc
chup man hinh) - hoan toan trong, khong don, khong mon, khong giao dich. Da
sao luu ra tep truoc khi xoa.

README (8.7.7)

Bo sung dung nhung buoc mot nguoi dung may trang se vap: tep SQLite khong nam
trong kho nen phai tao truoc khi migrate (thieu no thi migrate dung lai hoi,
kich ban tu dong se treo); CA_BUNDLE cho Windows (thieu la cURL error 60);
cach lay du lieu demo day du; bay .next dung chung giua build va dev. Sua so
bai kiem thu tu 29 thanh 195 va liet ke dung cac bo hien co.

+++ backend/app/Models/Order.php
@@ -13,10 +13,22 @@ class Order extends Model
         // trạng thái thanh toán + tiền mặt/thối.
         'invoice_code', 'payment_method', 'payment_status', 'cash_received', 'change_amount'];
 
+    /**
+     * MỌI trường tiền là SỐ NGUYÊN. Đồng Việt Nam không có đơn vị nhỏ hơn đồng, nên
+     * số thực ở đây chỉ mang lại rắc rối: hóa đơn hiện số lẻ mà không ai gõ vào, và
+     * phép cộng nhiều dòng tích lũy sai số nhị phân (0.1 + 0.2 ≠ 0.3).
+     *
+     * Trước đây `subtotal`, `discount_amount`, `total_amount` ép kiểu 'float' trong
+     * khi `cash_received`, `change_amount` đã là 'integer', và `Item::base_price`
+     * cũng vậy — cùng một đồng tiền mà hai kiểu khác nhau tùy chỗ đọc.
+     *
+     * Lưu ý cho người sửa sau: cast 'integer' của Laravel CẮT phần thập phân chứ
+     * không làm tròn. Nơi tính tiền phải `round()` trước khi ghi, đừng dựa vào cast.
+     */
     protected $casts = [
-        'subtotal' => 'float',
-        'discount_amount' => 'float',
-        'total_amount' => 'float',
+        'subtotal' => 'integer',
+        'discount_amount' => 'integer',
+        'total_amount' => 'integer',
         'cash_received' => 'integer',
         'change_amount' => 'integer',
     ];
